finish moving login over to top level views
